move db up a folder
